- 新增抽象流程实例详情策略基类，支持字段权限与按钮配置
-优化节点VO策略，增强审批人配置校验与去重逻辑
- 增加实体字段主键ID获取及条件构建方法
- 支持根据应用ID过滤角色与用户列表
-修复部分API接口参数与返回值类型不匹配问题- 完善安全访问拒绝处理器，提升权限控制日志记录
- 调整用户服务接口，支持昵称查询与租户用户统计-优化应用管理模块，支持按创建者筛选与用户头像展示- 禁用部分角色成员管理接口，待后续完善后重新启用
- 更新API请求过滤路径，适配平台、构建与运行时接口前缀

+++ onebase-module-system/onebase-module-system-core/src/main/java/com/cmsr/onebase/module/system/vo/corp/CorpReqVO.java
@@ -16,7 +16,7 @@ public class CorpReqVO {
     private String corpCode;
 
     @Schema(description = "行业类型", example = "1")
-    private Integer industryType;
+    private Long industryType;
 
     @Schema(description = "状态", example = "1")
     private Integer status;
